lib/token: split large file into multiple ones.
This is another nop: no changes to the code other than moving
functions around and invoking `go fmt` on all files.

Background: token.go is quite large, > 500 lines of code.
In preparation to add more asymmetric functions, splitting
the file into multiple ones.

Tested: all tests are still passing. Coverage is low, but
this change has not modified the code in any way.
