+++ lib/token/token_test.go
@@ -1,33 +1,11 @@
 package token
 
 import (
+	"context"
 	"github.com/stretchr/testify/assert"
 	"math/rand"
 	"testing"
-	"time"
-	"context"
 )
-
-func TestSimple(t *testing.T) {
-	rng := rand.New(rand.NewSource(1))
-	be, err := NewSymmetricEncoder(rng)
-	assert.Nil(t, be)
-	assert.NotNil(t, err)
-
-	be, err = NewSymmetricEncoder(rng, WithGeneratedSymmetricKey(12))
-	assert.Nil(t, be)
-	assert.NotNil(t, err)
-
-	be, err = NewSymmetricEncoder(rng, WithGeneratedSymmetricKey(128))
-	assert.NotNil(t, be)
-	assert.Nil(t, err)
-
-	data, err := be.Encode([]byte{1, 2, 3, 4})
-	assert.Nil(t, err)
-	_, original, err := be.Decode(context.Background(), data)
-	assert.Nil(t, err)
-	assert.Equal(t, []byte{1, 2, 3, 4}, original)
-}
 
 func TestTypeEncoder(t *testing.T) {
 	rng := rand.New(rand.NewSource(1))
@@ -48,32 +26,4 @@ func TestTypeEncoder(t *testing.T) {
 	_, err = te.Decode(context.Background(), data1, &result)
 	assert.Nil(t, err)
 	assert.Equal(t, "this is a string", result)
-}
-
-func TestTimeEncoder(t *testing.T) {
-	rng := rand.New(rand.NewSource(1))
-	be, err := NewSymmetricEncoder(rng, WithGeneratedSymmetricKey(128))
-	assert.NotNil(t, be)
-	assert.Nil(t, err)
-
-	ts := time.Now()
-	te := NewChainedEncoder(be, NewTimeEncoder(func() time.Time {
-		return ts
-	}, time.Second*5))
-	assert.NotNil(t, te)
-
-	data, err := te.Encode([]byte{0, 1, 2, 3, 4})
-	assert.Nil(t, err)
-	assert.NotNil(t, data)
-
-	ts = ts.Add(time.Second * 2)
-	_, arr, err := te.Decode(context.Background(), data)
-	assert.Nil(t, err)
-	assert.NotNil(t, arr)
-	assert.Equal(t, []byte{0, 1, 2, 3, 4}, arr)
-
-	// Now the timer has expired.
-	ts = ts.Add(time.Second * 3)
-	_, arr, err = te.Decode(context.Background(), data)
-	assert.NotNil(t, err)
 }
